🚀 feat:  公開と参加のViewを共通化
#93
