+++ lib/presentation/page/stamp_rally/component/stamp_rally_list_view.dart
@@ -2,12 +2,15 @@ import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../../../../application/stamp_rally/stamp_rally_service.dart';
 import '../../../../domain/repository/stamp_rally/entity/stamp_rally.dart';
 import '../../../component/cached_manager.dart';
-import '../../../component/dialog.dart';
-import '../../../component/thumbnail.dart';
 import '../../../router.dart';
+import 'spot_thumbnail.dart';
+import 'stamp_rally_button.dart';
+import 'stamp_rally_section.dart';
+
+/// 横方向のパディング
+const _horizontalPadding = 16.0;
 
 /// スタンプラリー詳細のリスト表示
 class StampRallyListView extends ConsumerStatefulWidget {
@@ -31,56 +34,169 @@ class _StampRallyListViewState extends ConsumerState<StampRallyListView>
   Widget build(BuildContext context) {
     super.build(context);
 
-    // Todo Figmaにあわせてデザイン実装
-    final stampRally = widget.stampRally;
-    return SingleChildScrollView(
-      child: Column(
-        children: [
-          ThumbnailImage(
-            imageUrl: stampRally.imageUrl,
+    return Stack(
+      children: [
+        // スタンプラリー画像
+        CachedNetworkImage(
+          imageUrl: widget.stampRally.imageUrl,
+          cacheManager: ref.watch(defaultCacheManagerProvider),
+        ),
+
+        // スタンプラリーの説明
+        _StampRallyExplanation(
+          stampRally: widget.stampRally,
+          topMargin: 240,
+        ),
+      ],
+    );
+  }
+}
+
+/// スタンプラリーの説明
+class _StampRallyExplanation extends StatelessWidget {
+  const _StampRallyExplanation({
+    required this.stampRally,
+    required this.topMargin,
+  });
+
+  final StampRally stampRally;
+  final double topMargin;
+
+  @override
+  Widget build(BuildContext context) {
+    return Stack(
+      children: [
+        Container(
+          decoration: BoxDecoration(
+            color: Theme.of(context).colorScheme.background,
+            borderRadius: BorderRadius.circular(16),
           ),
-          Text(stampRally.title),
-          Text('チェックポイント数：${stampRally.spots.length}'),
-          ElevatedButton(
-            onPressed: () => showDialog<void>(
-              context: context,
-              builder: (context) => ConfirmDialog(
-                message: '参加を完了しますか？',
-                onApproved: () async {
-                  await ref
-                      .read(stampRallyServiceProvider)
-                      .completeStampRally(stampRallyId: stampRally.id);
-                },
+          margin: EdgeInsets.only(top: topMargin),
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              Padding(
+                padding: const EdgeInsets.only(
+                  left: _horizontalPadding,
+                  right: StampRallyButton.buttonSize + 28,
+                  top: 16,
+                  bottom: 4,
+                ),
+                child: Text(
+                  stampRally.title,
+                  style: Theme.of(context).textTheme.titleLarge,
+                ),
+              ),
+              const Divider(
+                thickness: 1.5,
+                indent: _horizontalPadding,
+                endIndent: _horizontalPadding,
+              ),
+              Expanded(
+                child: SingleChildScrollView(
+                  padding: const EdgeInsets.symmetric(
+                    horizontal: _horizontalPadding,
+                  ),
+                  child: Column(
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    children: [
+                      _StampedCountText(
+                        stampRally: stampRally,
+                      ),
+                      StampRallySection(
+                        icon: Icons.location_on,
+                        title: 'エリア',
+                        body: stampRally.area,
+                      ),
+                      StampRallySection(
+                        icon: Icons.schedule,
+                        title: '所要時間',
+                        body: '${stampRally.requiredTime.toString()} 時間',
+                      ),
+                      StampRallySection(
+                        icon: Icons.summarize,
+                        title: '概要',
+                        body: stampRally.summary,
+                      ),
+                      GridView.builder(
+                        shrinkWrap: true,
+                        physics: const NeverScrollableScrollPhysics(),
+                        gridDelegate:
+                            const SliverGridDelegateWithFixedCrossAxisCount(
+                          crossAxisCount: 2,
+                        ),
+                        itemCount: stampRally.spots.length,
+                        itemBuilder: (context, index) {
+                          final spot = stampRally.spots[index];
+                          return InkWell(
+                            onTap: () async {
+                              PublicSpotViewRoute.fromSpot(spot).go(context);
+                            },
+                            child: SpotThumbnail(
+                              spot: spot,
+                            ),
+                          );
+                        },
+                      ),
+                    ],
+                  ),
+                ),
+              )
+            ],
+          ),
+        ),
+
+        // 参加／完了／中断ボタン
+        Container(
+          alignment: Alignment.topRight,
+          margin: EdgeInsets.only(
+            top: topMargin - StampRallyButton.buttonSize / 2,
+            right: 20,
+          ),
+          child: StampRallyControlButton(
+            stampRally: stampRally,
+          ),
+        ),
+      ],
+    );
+  }
+}
+
+/// スタンプ数
+///
+/// 参加スタンプラリーのみ表示する
+class _StampedCountText extends StatelessWidget {
+  const _StampedCountText({
+    required this.stampRally,
+  });
+
+  final StampRally stampRally;
+
+  @override
+  Widget build(BuildContext context) {
+    if (!stampRally.isEntry) {
+      return const SizedBox();
+    }
+
+    return Padding(
+      padding: const EdgeInsets.only(bottom: 8),
+      child: DefaultTextStyle.merge(
+        style: Theme.of(context).textTheme.headline3?.copyWith(
+              color: Theme.of(context).colorScheme.onBackground,
+            ),
+        child: Row(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Text(
+              '${stampRally.stampedCount}',
+              style: TextStyle(
+                color: Theme.of(context).colorScheme.primaryContainer,
               ),
             ),
-            child: const Text('参加完了'),
-          ),
-          ElevatedButton(
-            onPressed: () => showDialog<void>(
-              context: context,
-              builder: (context) => ConfirmDialog(
-                message: '本当に参加を中断しますか？',
-                onApproved: () async {
-                  await ref
-                      .read(stampRallyServiceProvider)
-                      .withdrawStampRally(stampRallyId: stampRally.id);
-                },
-              ),
-            ),
-            child: const Text('参加中断'),
-          ),
-          ...stampRally.spots.map(
-            (spot) => InkWell(
-              onTap: () async {
-                EntrySpotViewRoute.fromSpot(spot).go(context);
-              },
-              child: CachedNetworkImage(
-                imageUrl: spot.imageUrl,
-                cacheManager: ref.watch(defaultCacheManagerProvider),
-              ),
-            ),
-          ),
-        ],
+            const Text(' / '),
+            Text('${stampRally.spots.length}'),
+          ],
+        ),
       ),
     );
   }
